Move util to its own folder
