+++ ddk/src/transport/nostr/messages.rs
@@ -1,6 +1,6 @@
 use crate::nostr::{DLC_MESSAGE_KIND, ORACLE_ANNOUNCMENT_KIND, ORACLE_ATTESTATION_KIND};
 use crate::transport::nostr::nostr_to_bitcoin_pubkey;
-use crate::util::message_variant_name;
+use crate::util::ser::message_variant_name;
 use dlc::secp256k1_zkp::PublicKey as SecpPublicKey;
 use dlc_messages::message_handler::read_dlc_message;
 use dlc_messages::{Message, WireMessage};
